+++ extensions/sql-database-projects/CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to the SQL Database Projects extension will be documented in
 
 ## [Unreleased]
 
+## [1.5.7] - 2026-02-27
+
+- Added five new SQL object templates: Schema, Table-Valued Function, Trigger, Database Trigger, and Sequence.
+- Fixed an issue where the SQL project build task was being created at the project level instead of the workspace level.
+
+
 ## [1.5.6] - 2026-01-28
 
 - Added a new 'Target platform' selector when creating an Azure SQL database project.